requirements.txt
django
Pillow 
matplotlib
google-generativeai
python-dotenv
numpy
requests